+++ api/src/OpenTelemetry/Resource.hs
@@ -160,8 +160,9 @@ class ToResource a where
 
 
 class MaterializeResource schema where
-  -- | Convert resource fields into a version that discharges the schema from the
-  -- type level to the runtime level.
+  {- | Convert resource fields into a version that discharges the schema from the
+  type level to the runtime level.
+  -}
   materializeResources :: Resource schema -> MaterializedResources